+codestyle 4

+++ tasks/mpi/kovalev_k_bubble_sort_oddeven_transposition/func_tests/main.cpp
@@ -40,7 +40,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_mpi, Test_NoViol_300_int) {
   tmpTaskPar.post_processing();
   if (world.rank() == 0) {
     int count_viol = 0;
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
       if (out[i] != in[i]) count_viol++;
     }
     ASSERT_EQ(count_viol, 0);
@@ -51,7 +51,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_mpi, Test_300_int) {
   const size_t length = 300;
   std::srand(std::time(nullptr));
   std::vector<int> in(length);
-  for (int i = 0; i < length; i++) in[i] = rand() * pow(-1, rand());
+  for (size_t i = 0; i < length; i++) in[i] = rand() * pow(-1, rand());
   std::vector<int> out(length);
   boost::mpi::communicator world;
   std::shared_ptr<ppc::core::TaskData> tmpPar = std::make_shared<ppc::core::TaskData>();
@@ -69,7 +69,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_mpi, Test_300_int) {
   if (world.rank() == 0) {
     std::sort(in.begin(), in.end(), [](int a, int b) { return a < b; });
     int count_viol = 0;
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
       if (out[i] != in[i]) count_viol++;
     }
     ASSERT_EQ(count_viol, 0);
@@ -80,7 +80,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_mpi, Test_30000_int) {
   const size_t length = 30000;
   std::srand(std::time(nullptr));
   std::vector<int> in(length);
-  for (int i = 0; i < length; i++) in[i] = rand() * pow(-1, rand());
+  for (size_t i = 0; i < length; i++) in[i] = rand() * pow(-1, rand());
   std::vector<int> out(length);
   boost::mpi::communicator world;
   std::shared_ptr<ppc::core::TaskData> tmpPar = std::make_shared<ppc::core::TaskData>();
@@ -98,7 +98,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_mpi, Test_30000_int) {
   if (world.rank() == 0) {
     std::sort(in.begin(), in.end(), [](int a, int b) { return a < b; });
     int count_viol = 0;
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
       if (out[i] != in[i]) count_viol++;
     }
     ASSERT_EQ(count_viol, 0);
@@ -129,7 +129,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_mpi, Test_300_double) {
   if (world.rank() == 0) {
     std::sort(in.begin(), in.end(), [](double a, double b) { return a < b; });
     int count_viol = 0;
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
       if (out[i] != in[i]) count_viol++;
     }
     ASSERT_EQ(count_viol, 0);
@@ -160,7 +160,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_mpi, Test_30000_double) {
   if (world.rank() == 0) {
     std::sort(in.begin(), in.end(), [](double a, double b) { return a < b; });
     int count_viol = 0;
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
       if (out[i] != in[i]) count_viol++;
     }
     ASSERT_EQ(count_viol, 0);
