+++ tasks/seq/kovalev_k_bubble_sort_oddeven_transposition/func_tests/main.cpp
@@ -32,7 +32,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_seq, Test_No_viol_10_int) {
   tmpTaskSeq.run();
   tmpTaskSeq.post_processing();
   int count_viol = 0;
-  for (int i = 0; i < length; i++) {
+  for (size_t i = 0; i < length; i++) {
     if (out[i] != in[i]) count_viol++;
   }
   ASSERT_EQ(count_viol, 0);
@@ -56,7 +56,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_seq, Test_1000_int) {
   tmpTaskSeq.post_processing();
   std::sort(in.begin(), in.end(), [](int a, int b) { return a < b; });
   int count_viol = 0;
-  for (int i = 0; i < length; i++) {
+  for (size_t i = 0; i < length; i++) {
     if (out[i] != in[i]) count_viol++;
   }
   ASSERT_EQ(count_viol, 0);
@@ -80,7 +80,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_seq, Test_10000_int) {
   tmpTaskSeq.post_processing();
   std::sort(in.begin(), in.end(), [](int a, int b) { return a < b; });
   int count_viol = 0;
-  for (int i = 0; i < length; i++) {
+  for (size_t i = 0; i < length; i++) {
     if (out[i] != in[i]) count_viol++;
   }
   ASSERT_EQ(count_viol, 0);
@@ -103,7 +103,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_seq, Test_No_viol_10_double) {
   tmpTaskSeq.run();
   tmpTaskSeq.post_processing();
   int count_viol = 0;
-  for (int i = 0; i < length; i++) {
+  for (size_t i = 0; i < length; i++) {
     if (out[i] != in[i]) count_viol++;
   }
   ASSERT_EQ(count_viol, 0);
@@ -129,7 +129,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_seq, Test_1000_double) {
   tmpTaskSeq.post_processing();
   std::sort(in.begin(), in.end(), [](double a, double b) { return a < b; });
   int count_viol = 0;
-  for (int i = 0; i < length; i++) {
+  for (size_t i = 0; i < length; i++) {
     if (out[i] != in[i]) count_viol++;
   }
   ASSERT_EQ(count_viol, 0);
@@ -155,7 +155,7 @@ TEST(kovalev_k_bubble_sort_oddeven_transposition_seq, Test_10000_double) {
   tmpTaskSeq.post_processing();
   std::sort(in.begin(), in.end(), [](double a, double b) { return a < b; });
   int count_viol = 0;
-  for (int i = 0; i < length; i++) {
+  for (size_t i = 0; i < length; i++) {
     if (out[i] != in[i]) count_viol++;
   }
   ASSERT_EQ(count_viol, 0);
